testproject/requirements.txt
Django==2.2.10
djangorestframework==3.13.1
drf-nested-routers==0.93.4